FileUtils : Was not correctly handling the case where a File object became null, it would previously have thrown a NPE which was erroneously caught this and declared the delete to have failed. If there is nothing to delete (signified by the Null File object) then the delete should pass.
git-svn-id: https://svn.eu.apache.org/repos/asf/qpid/trunk@750946 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/java/common/src/main/java/org/apache/qpid/util/FileUtils.java
@@ -246,18 +246,19 @@ public class FileUtils
      {
          boolean success = true;
 
+         // If we have nothing to delete then it must be ok to say it was deleted.
+         if (file == null)
+         {
+             return true;
+         }
+
          if (file.isDirectory())
          {
              if (recursive)
              {
-                 try{
-                     for (File subFile : file.listFiles())
-                     {
-                         success = delete(subFile, true) & success ;
-                     }
-                 }catch (NullPointerException npe)
+                 for (File subFile : file.listFiles())
                  {
-                     success = false;
+                     success = delete(subFile, true) & success ;
                  }
 
                  return success && file.delete();
